rename: 컨벤션에 따라 컴포넌트명 파스칼 케이스로 변경
